chore: correcoes de importacoes e migracoes de pastas das telas de pacientes
