+++ apps/agendamento/src/app/pessoa/cadastro/adicional/page.tsx
@@ -12,8 +12,8 @@ import { Input } from "@/components/ui/input";
 import {
   MembersRegisterStep,
   useMembersRegisterContext,
-} from "@/hooks/use-members-register-context";
-import { Additionals } from "@/schemas/member-schemas";
+} from "@/../../management-app/src/hooks/use-members-register-context";
+import { Additionals } from "@/app/schemas/member-schemas";
 import { zodResolver } from "@hookform/resolvers/zod";
 import React from "react";
 import { useForm } from "react-hook-form";
