+++ apps/agendamento/src/app/pessoa/cadastro/endereco/page.tsx
@@ -12,9 +12,9 @@ import { Input } from "@/components/ui/input";
 import {
   MembersRegisterStep,
   useMembersRegisterContext,
-} from "@/hooks/use-members-register-context";
+} from "@/../../management-app/src/hooks/use-members-register-context";
 import { formatCEP } from "@/lib/formats";
-import { Address } from "@/schemas/member-schemas";
+import { Address } from "@/app/schemas/member-schemas";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import z from "zod";
