+++ apps/agendamento/src/app/pessoa/cadastro/layout.tsx
@@ -1,5 +1,5 @@
 import Image from "@/app/assets/background_image.jpg";
-import { MembersRegisterProvider } from "@/hooks/use-members-register-context";
+import { MembersRegisterProvider } from "@/../../management-app/src/hooks/use-members-register-context";
 import { PageOrchestrator } from "./orchestrator";
 import { Nunito } from "next/font/google";
