+++ apps/agendamento/src/app/transtornos/new/page.tsx
@@ -1,58 +1,38 @@
 "use client";
 
-import { useRouter, useParams } from "next/navigation";
+import { useRouter } from "next/navigation";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { updateTranstornoSchema, UpdateTranstornoDTO } from "@/schemas/transtornosSchema";
+import { createTranstornoSchema, CreateTranstornoDTO } from "@/app/schemas/transtornosSchema";
 import { toast } from "react-toastify";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { useEffect } from "react";
-import { Loader2, ArrowLeft } from "lucide-react";
+import { Label } from "@/components/ui/label"; 
+import { ArrowLeft } from "lucide-react";
 
-export default function EditTranstornoPage() {
+export default function NewTranstornoPage() {
   const router = useRouter();
-  const params = useParams();
-  const { id } = params;
-
   const {
     register,
     handleSubmit,
-    setValue,
     formState: { errors, isSubmitting },
-  } = useForm<UpdateTranstornoDTO>({
-    resolver: zodResolver(updateTranstornoSchema),
+  } = useForm<CreateTranstornoDTO>({
+    resolver: zodResolver(createTranstornoSchema),
   });
 
-  useEffect(() => {
-    if (id) {
-      const fetchTranstorno = async () => {
-        try {
-          const response = await fetch(`/api/transtornos/${id}`);
-          if (!response.ok) throw new Error("Transtorno não encontrado.");
-          const data = await response.json();
-          setValue("name", data.name);
-        } catch (error: any) {
-          toast.error(error.message);
-          router.push("/transtornos");
-        }
-      };
-      fetchTranstorno();
-    }
-  }, [id, setValue, router]);
-
-  const onSubmit = async (data: UpdateTranstornoDTO) => {
+  const onSubmit = async (data: CreateTranstornoDTO) => {
     try {
-      const response = await fetch(`/api/transtornos/${id}`, {
-        method: "PUT",
+      const response = await fetch("/api/transtornos", {
+        method: "POST",
         headers: { "Content-Type": "application/json" },
         body: JSON.stringify(data),
       });
+
       if (!response.ok) {
-        throw new Error("Falha ao atualizar transtorno.");
+        throw new Error("Falha ao criar transtorno.");
       }
-      toast.success("Transtorno atualizado com sucesso!");
+
+      toast.success("Transtorno criado com sucesso!");
       router.push("/transtornos");
       router.refresh();
     } catch (error: any) {
@@ -60,29 +40,20 @@ export default function EditTranstornoPage() {
     }
   };
 
-  if (!id) {
-    return (
-      <div className="flex justify-center items-center p-10">
-        <Loader2 className="h-8 w-8 animate-spin text-gray-500" />
-      </div>
-    );
-  }
-
-
   return (
-    <div className="!bg-slate-100 min-h-screen">
+    <div className="flex-1 !bg-slate-100 min-h-screen">
       <main className="container mx-auto p-4 md:p-6">
         <div className="bg-white rounded-xl shadow-md border-2 p-6 mb-4">
-          <Button
-            variant="ghost"
-            onClick={() => router.back()}
+          <Button 
+            variant="ghost" 
+            onClick={() => router.back()} 
             className="mb-4 text-sm text-[#003B93] hover:bg-blue-50"
           >
             <ArrowLeft className="h-4 w-4 mr-2" />
             Voltar
           </Button>
           
-          <h1 className="text-2xl font-bold mb-6 text-[#003B93]">Editar Transtorno</h1>
+          <h1 className="text-2xl font-bold mb-6 text-[#003B93]">Novo Transtorno</h1>
           
           <form onSubmit={handleSubmit(onSubmit)} className="space-y-4">
             <div>
@@ -93,6 +64,7 @@ export default function EditTranstornoPage() {
                 id="name"
                 {...register("name")}
                 className="mt-1 block w-full"
+                placeholder="Ex: TDAH"
               />
               {errors.name && (
                 <p className="mt-1 text-sm text-red-600">{errors.name.message}</p>
@@ -105,10 +77,10 @@ export default function EditTranstornoPage() {
               </Button>
               <Button 
                 type="submit" 
-                disabled={isSubmitting}
+                disabled={isSubmitting} 
                 className="!bg-[#0D4F97] !hover:bg-[#0b427d] text-white"
               >
-                {isSubmitting ? "Atualizando..." : "Atualizar"}
+                {isSubmitting ? "Salvando..." : "Salvar"}
               </Button>
             </div>
           </form>
