+++ apps/agendamento/src/app/visualization-patients/page.tsx
@@ -77,7 +77,7 @@ export default function PatientsAndStudentsScreen() {
   };
 
   return (
-    <div className="!bg-slate-100 min-h-screen">
+    <div className="flex-1 !bg-slate-100 min-h-screen">
       <main className="container mx-auto p-4 md:p-6">
         <div className="bg-white rounded-xl shadow-md border-2 p-6 mb-4">
           <SearchFilters
